+++ src/components/Welcome/index.ts
@@ -1,0 +1,2 @@
+export { Welcome } from './Welcome'
+export { markWelcomeSeen, welcomeSeen } from './welcomeSeen'